+++ src/app/layout/Footer.tsx
@@ -1,4 +1,4 @@
-import { FooterContainer } from './layout.style';
+import { FooterContainer } from '@/app/layout';
 
 export const Footer = () => {
   return <FooterContainer>Footer</FooterContainer>;
