+++ src/app/layout/Header.tsx
@@ -1,4 +1,4 @@
-import { HeaderContainer } from './layout.style';
+import { HeaderContainer } from '@/app/layout';
 
 export const Header = () => {
   return <HeaderContainer>Header</HeaderContainer>;
