+++ src/app/layout/index.ts
@@ -1,1 +1,2 @@
 export { Layout } from './Layout';
+export { HeaderContainer, FooterContainer, Container } from '@/app/layout/layout.style';
